Fix megaface eval as only one out is returned
This was changed recently (but before this was pushed) to have only one output. So pulling first element now breaks it.

I had to make numerous changes to get this to work with Python 3 but I believe this change would break in both.
I also have an issue where it finds the GPU free memory size in MB instead of in GB but haven't investigated.